No me anda el atacar cuando apreto la tecla A pero si sale por consola que me la toma. A ustedes les anda?

+++ BombitaTp2/src/main/ControlJuego.java
@@ -123,7 +123,7 @@ public class ControlJuego {
 		return bombita.obtenerVida();
 	}
 	
-	public float obtenerVelocidadArea() {
+	public int obtenerVelocidadArea() {
 		return bombita.getVelocidad();
 	}
